professeur cours

+++ professeurDashboard/deleteCours.php
@@ -4,15 +4,24 @@ require_once('config.php');
 if(isset($_GET['id'])) {
     $coursId = $_GET['id'];
 
-    $sqlDeleteCours = "DELETE FROM cours WHERE coursId = '$coursId'";
-    $resultCours = mysqli_query($conn, $sqlDeleteCours);
+    $sqlDeleteChapitres = "DELETE FROM chapitre WHERE coursId = '$coursId'";
+    $resultChapitres = mysqli_query($conn, $sqlDeleteChapitres);
 
-    if($resultCours) {
-        header("Location: cours.php?cbe=cbe");
+    if($resultChapitres) {
+        // Ahora que se han eliminado los capítulos asociados, procedemos a eliminar el curso
+        $sqlDeleteCours = "DELETE FROM cours WHERE coursId = '$coursId'";
+        $resultCours = mysqli_query($conn, $sqlDeleteCours);
+
+        if($resultCours) {
+            header("Location: cours.php?cbe=cbe");
+        } else {
+            echo "Error al intentar eliminar el curso.";
+        }
     } else {
-        echo "Error al intentar eliminar el curso.";
+        echo "Error al intentar eliminar los capítulos del curso.";
     }
 } else {
     echo "ID de curso no proporcionado.";
 }
 ?>
+
